signal: how kernel continues after SIGKILL
According to thoery, the process should get KILLED, no lines should get executed after another process delivered it SIGKILL.But if that process is in kernel space when signal is delivered, then the user space lines are not executed (after the system call which resulted in the process go into kernel space) but kernbel space lines after "wait_event_interruptible_timeout" are getting executed. How?. Even that should be aborted according to theory.